+++ ignition/modules/IsoniaProtocolV01.ts
@@ -6,7 +6,7 @@ export default buildModule("IsoniaProtocolV01Module", (m) => {
   const govCore = m.contract("GovCore");
   const demoTarget = m.contract("DemoTarget", [deployer]);
   const demoVotesToken = m.contract("IsoDemoVotesToken", [deployer]);
-  const govProposals = m.contract("GovProposals", [govCore, demoTarget]);
+  const govProposals = m.contract("GovProposals", [govCore]);
 
   m.call(demoTarget, "setGovProposals", [govProposals]);